Don't mess with Django's default rendering
Prior to this patch, Django admin would try to render the widget
templates that we instanciate with "from_boundfield" like normal
templates.

At the same time, we don't really want our form rendering to mess with
Django, it should only be active when being rendered as a component,
let form template rendering work as it used too outside of ryzom
(to_html calls).